Fix main layout and fix evergreen issue
