+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "@auth0/nextjs-auth0": "^3.5.0",
     "dompurify": "^3.0.3",
-    "evergreen-ui": "^6.1.0",
+    "evergreen-ui": "^7.1.9",
     "next": "14.2.15",
     "next-pwa": "^5.6.0",
     "react": "^18.3.1",
